% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.mgf.formatted.spectra.R
\name{export.mgf.formatted.spectra}
\alias{export.mgf.formatted.spectra}
\title{export.mgf.formatted.spectra}
\usage{
export.mgf.formatted.spectra(ramclustObj = NULL, output.dir = NULL)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{output.dir}{character: path to write MGF formatted spectra to a new directory called 'spectra' is created (if it does not exist), and within it a directory called 'mgf'.  all spectra files are written to the 'spectra/ms' directory within the specified directory, or the working directory, if not specified.}
}
\value{
nothing - files written to disk
}
\description{
export.mgf.formatted.spectra
}
\details{
This function exports spectra in .mgf format, suitable for importing into GNPS, NIST MSsearch, etc.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{xcms}
