% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiora.mgf.to.spectra.R
\name{fiora.mgf.to.spectra}
\alias{fiora.mgf.to.spectra}
\title{fiora.mgf.to.spectra}
\usage{
fiora.mgf.to.spectra(
  mgf.dir = NULL,
  split.by.polarity = TRUE,
  out.dir = NULL,
  return.library = TRUE
)
}
\arguments{
\item{mgf.dir}{character file path.  directory containing fiora output .mgf files.}

\item{split.by.polarity}{logical.  if true, spectra are split into positive ('1') and negative ('0') polarities before saving. default = TRUE.}

\item{out.dir}{character. full file path to save 'Spectra' object(s) to.}

\item{return.library}{logical.  If true, return library to R console.}
}
\value{
spectra object.  File(s) also sent to output directory 'out.dir'.
}
\description{
imports fiora .mgf files and formats, with naming scheme from 'pubchem.bio.to.fiora', into R as a 'Spectra' object, saves as new .Rdata object on disk
}
\details{
takes a pubchem.bio object (data.table) and writes out input files for fiora. optionally splits into multiple files for RAM management, writes out text file of commands suitable for running batch processing.
}
\author{
Corey Broeckling
}
