% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc.ei.create.spectra.object.R
\name{gc.ei.create.spectra}
\alias{gc.ei.create.spectra}
\title{gc.ei.create.spectra}
\usage{
gc.ei.create.spectra(ramclustObj = NULL, round.digits = 0, mult = 0.9988)
}
\arguments{
\item{ramclustObj}{ramclustR object input.}

\item{round.digits}{how many digits to round mz value to.  for nist search, best to use '0'.  for other GC-EI databases, you could use a larger value, but then you should set 'mult = 1'}

\item{mult}{the multiplier for m/z to ensure that rounding results in the proper 'floor' mass.  NIST recommends 0.9988 to account for CH2 mass defect increase with molecular weight.}
}
\value{
$ms1.spectrum - list of all spectra for all clusters.

$feature.table - organized table with one row per feature.
}
\description{
Cluster annotation function: inference of 'M' - molecular weight of the compound giving rise to each spectrum - using the InterpretMSSpectrum::findMain function
}
\details{
returns ramclustObj with GC-EI spectra organized into a 'Spectra' object for searching.
}
\references{
Jaeger C, ... Lisec J. Compound annotation in liquid chromatography/high-resolution mass spectrometry based metabolomics: robust adduct ion determination as a prerequisite to structure prediction in electrospray ionization mass spectra. Rapid Commun Mass Spectrom. 2017 Aug 15;31(15):1261-1266. doi: 10.1002/rcm.7905. PubMed PMID: 28499062.

Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Ganna A, Layer M, Brown K, Sutton B, Ingelsson E, Peers G, Prenni JE. Enabling Efficient and Confident Annotation of LC-MS Metabolomics Data through MS1 Spectrum and Time Prediction. Anal Chem. 2016 Sep 20;88(18):9226-34. doi: 10.1021/acs.analchem.6b02479. Epub 2016 Sep 8. PubMed PMID: 7560453.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{findMain}
\concept{interpretMSSpectrum}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
