% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.arcbio.R.R
\name{setup.arcbio.R}
\alias{setup.arcbio.R}
\title{setup.arcbio.R}
\usage{
setup.arcbio.R()
}
\value{
nothing
}
\description{
just a convenience function to set up all packages needed for routine metabolomics work at CSU ARC-BIO
}
\details{
can use "source('https://raw.githubusercontent.com/cbroeckl/csu.arcbio.tools/master/R/setup.arcbio.R.R'); setup.arcbio.R()" to run
}
\author{
Corey Broeckling
}
