% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiora.mgf.to.spectra.R,
%   R/pubchem.bio.to.fiora.R
\name{pubchem.bio.to.fiora}
\alias{pubchem.bio.to.fiora}
\title{fiora.mgf.to.spectra}
\usage{
pubchem.bio.to.fiora(
  pubchem.bio.object = NULL,
  max.cmpds.per.file = 1000,
  param.table = data.frame(Precursor_type = "[M+H]+", CE = 20, Instrument_type = "HCD"),
  min_prob = 0.01,
  annotation = TRUE,
  rm.charged = FALSE,
  mw.range = c(50, 1200),
  out.dir = NULL
)

pubchem.bio.to.fiora(
  pubchem.bio.object = NULL,
  max.cmpds.per.file = 1000,
  param.table = data.frame(Precursor_type = "[M+H]+", CE = 20, Instrument_type = "HCD"),
  min_prob = 0.01,
  annotation = TRUE,
  rm.charged = FALSE,
  mw.range = c(50, 1200),
  out.dir = NULL
)
}
\arguments{
\item{pubchem.bio.object}{R data.table, generally produced by build.pubchem.bio.}

\item{max.cmpds.per.file}{integer. default = 1000.  use more if you have lots of RAM, less if you have little.}

\item{param.table}{data frame. should contain headers 'Precursor_type', 'CE', and 'Instrument_type'.  Each row becomes a parameter set to predict. See Fiora documentation for valid options.}

\item{min_prob}{numeric. fiora option for minimum probability for a fragment ion to be reported. default = 0.001 in Fiora, 0.01 in this function.}

\item{annotation}{logical. should fragment ion structures be reported in output mgf format? default = TRUE.}

\item{rm.charged}{logical. should charged molecules be removed?  default = FALSE.}

\item{mw.range}{numeric vector, length two.  c(50, 1200) by default. only compounds in this range are exported.}

\item{out.dir}{path.  full path to directory files are to be written to. Directory must already exist.}

\item{mgf.dir}{character file path.  directory containing fiora output .mgf files.}

\item{split.by.polarity}{logical.  if true, spectra are split into positive ('1') and negative ('0') polarities before saving. default = TRUE.}
}
\value{
spectra object.  File(s) also sent to output directory 'out.dir'.

data.frame of Fiora input parameters.  Files also sent to output directory 'out.dir'.
}
\description{
imports fiora .mgf files and formats, with naming scheme from 'pubchem.bio.to.fiora', into R as a 'Spectra' object, saves as new .Rdata object on disk

takes a pubchem.bio object (data.table) and writes out input files for fiora.
}
\details{
takes a pubchem.bio object (data.table) and writes out input files for fiora. optionally splits into multiple files for RAM management, writes out text file of commands suitable for running batch processing.

takes a pubchem.bio object (data.table) and writes out input files for fiora. optionally splits into multiple files for RAM management, writes out text file of commands suitable for running batch processing.
}
\author{
Corey Broeckling
}
